+++ x/posts/internal/types/post.go
@@ -172,7 +172,7 @@ func (p Post) Validate() error {
 	}
 
 	if len(strings.TrimSpace(p.Message)) == 0 && len(p.Medias) == 0 {
-		return fmt.Errorf("post message and medias can't be both empty")
+		return fmt.Errorf("post message or medias required, they cannot be both empty")
 	}
 
 	if len(p.Message) > MaxPostMessageLength {