+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Version 0.4.0
 ## Changes
 - Improved `alias.go` files (#103)
+- Added the support for posting empty-message posts with medias (#110)
 - Implemented the support for hashtags in posts (#96)
+
+## Bug fixes
+- Fixed a bug inside the migration procedure of the `magpie` module (#106)
 
 # Version 0.3.0
 ## Changes